[Interpreter] This patch adds support for variable sized argument lists
Internal: Support for X86-64 va_list was added using the same technique as
klee, the arguments are packed into an interpreter allocated object and a
pointer to this array is stored in va_list (this is equivalent to putting all
parameters onto the stack instead of in registers).

External: support for using libffi var arg functionality was added, this
requires passing a list of argument Types extracted from the caller
to callExternalFunction.

Not supported: passing a va_list object from within the interpreter
to an external lib call.